Add support for vendor extensions to CodegenModel/Property
Fix issue with modelFileTemplates supporting the notion of
multiple templates per model, but not providing a way to
actually output multiple files per model. Added a
toModelFilename() call which takes a template filename
argument to allow subclasses to special case the output
filename.

This is patch to support a model generator extension that
generates @AutoValue models as well as persistence entities
as separate files.